Addition of Steam Scraper
Steam scraper added in here. From Alex's original version, added a pipeline that outputs the files to a specific folder, as opposed to using scrapyd's standard output. The reason that this was done was so that we can split the file into multiple parts as opposed to one large file.